author backend işlemleri tamamlandı.Blog backend işlemleri beklemede.

+++ Core/CarBookNetCore.Application/Features/CQRS/Handlers/CarHandlers/RemoveCarCommandHandler.cs
@@ -20,7 +20,10 @@ namespace CarBookNetCore.Application.Features.CQRS.Handlers.CarHandlers
         public async Task Handle(RemoveCarCommand command)
         {
             var value = await _repository.GetByIdAsync(command.Id);
-            await _repository.RemoveAsync(value);
+            if (value != null)
+            {
+                await _repository.RemoveAsync(value);
+            }
         }
     }
 }
